+++ src/controllers/diet/food-items/food-item-controller.controller.ts
@@ -1,11 +1,11 @@
 import {Filter, repository} from '@loopback/repository';
 import {get, getModelSchemaRef, param} from '@loopback/rest';
-import {FoodGenusNutrientsPivot, FoodGroupItems} from '../models';
+import {FoodGenusNutrientsPivot, FoodGroupItems} from '../../../models';
 import {
   FoodGenusNutrientsPivotRepository,
   FoodGroupItemsRepository,
-} from '../repositories';
-import {StandardJsonResponse} from './standardJsonResponse';
+} from '../../../repositories';
+import {StandardJsonResponse} from '../../standardJsonResponse';
 
 export class FoodItemControllerController {
   constructor(
@@ -18,7 +18,7 @@ export class FoodItemControllerController {
   @get('/diet/food-group', {
     summary: 'Get Food Groups and Food Items',
     description: 'Return a list of food groups and their associated fooditems',
-    tags: ['diet'],
+    // tags: ['diet'],
     responses: {
       '200': {
         description: 'Array of FoodGroupItems model instances',
@@ -47,7 +47,7 @@ export class FoodItemControllerController {
 
   @get('/diet/composition', {
     summary: 'composition',
-    tags: ['diet'],
+    // tags: ['diet'],
     responses: {
       '200': {
         description: 'Array of FoodGenusNutrientsPivot model instances',