+++ src/enhancers/dapperdoxSpecEnhancer.ts
@@ -46,7 +46,7 @@ export class DapperdoxSpecEnhancer implements OASEnhancer {
     spec.info.title = 'Core Data API';
 
     spec.host = 'localhost:3000';
-    spec.schemes = ['https'];
+    //spec.schemes = ['https'];
 
     spec.securityDefinitions = {
       apiKey: {
@@ -62,6 +62,10 @@ export class DapperdoxSpecEnhancer implements OASEnhancer {
         {
           name: 'diet',
           description: 'Diet data',
+        },
+        {
+          name: 'biomarker',
+          description: 'Biomarker data',
         },
         {
           name: 'projections',